+++ flocx_ui/api/flocx_rest_api.py
@@ -64,6 +64,17 @@ class Bids(generic.View):
         bids = flocx_market.bid_list(request)
         return bids
 
+    @rest_utils.ajax()
+    def post(self, request):
+        """Create a bid
+
+        :param request: The request passed into the ajax decorator
+        :return: The created bid
+        """
+        bid_data = request.body
+        bid = flocx_market.bid_create(request, bid_data)
+        return bid
+
 @urls.register
 class Contracts(generic.View):