+++ flocx_ui/static/dashboard/project/flocx/bid-list/bid-list.controller.js
@@ -6,15 +6,28 @@
     .controller('BidListController', BidListController);
 
   BidListController.$inject = [
-    'horizon.app.core.openstack-service-api.flocx'
+    'horizon.app.core.openstack-service-api.flocx',
+    'horizon.dashboard.project.flocx.create-bid.service'
   ];
 
-  function BidListController(flocx) {
+  function BidListController(flocx, createBidService) {
     var ctrl = this;
 
+    ctrl.createBid = createBid;
     ctrl.bids = [];
 
     init();
+
+    /**
+     * @description Open a modal that guides a user through creating a bid
+     *
+     * @returns {void}
+     */
+    function createBid () {
+      // Open the modal
+      createBidService.createBid()
+        .then(init); // Refresh the bids
+    }
 
     function init () {
       retrieveBids();