+++ flocx_ui/static/dashboard/project/flocx/flocxApi.service.js
@@ -20,6 +20,7 @@
       getOffers: getOffers,
       createOffer: createOffer,
       getBids: getBids,
+      createBid: createBid,
       getContracts: getContracts
     };
 
@@ -54,6 +55,18 @@
      */
     function getBids() {
       return apiService.get('/api/flocx/bid/')
+        .then(function (response) {
+          return response.data;
+        });
+    }
+
+    /**
+     * @description Create a bid
+     * @param {*} bid The bid data to be used when creating the bid
+     * @returns {promise} Promise containing the created offer
+     */
+    function createBid(bid) {
+      return apiService.post('/api/flocx/bid/', bid)
         .then(function (response) {
           return response.data;
         });